+++ packages/ice/src/service/serverCompiler.ts
@@ -154,6 +154,12 @@ export function createServerCompiler(options: Options) {
       define,
       absWorkingDir: rootDir,
       external: Object.keys(externals),
+      banner: customBuildOptions.platform === 'node' && server?.format !== 'cjs'
+        ? {
+            // See https://github.com/evanw/esbuild/issues/1921#issuecomment-1152991694
+            js: 'import { createRequire } from \'module\';const require = createRequire(import.meta.url);',
+          }
+        : undefined,
       plugins: [
         ...(customBuildOptions.plugins || []),
         emptyCSSPlugin(),